+++ lib/graphql_rails/router/query_route.rb
@@ -14,7 +14,7 @@ module GraphqlRails
         false
       end
 
-      def subscription?
+      def event?
         false
       end
     end